add fileSrc to index, move suffixArrayCompression out of metadata, add alloc, dealloc, and helper functions
Index now stores the fileSrc, to keep an open index coupled to the file 
that it represents.
suffixArrayCompressionRatio isn't really a metadata field, so it makes 
more sense to belong in the AwFmIndex struct.

This commit also adds AwFmSearchRange, to represent a (start pointer - 
end pointer) range in the BWT, helpful for returning kmer suffix matches 
from functions.