change input types to uint8_t, add transform from vector letter to freq index letter.
The input types were changed to uint8_t because the C standard doesn't 
define whether char is signed or not, so it ~might~ index with a 
negative value, causing weird shit. This should not modify the input 
characters at all, just how they're used to index.

In order to do the backtrace to determine database sequence position, we 
need a function to convert letters in the AwFmBlock vectors back to 
frequency-indexed.