remove bidirectional index from index data structure
this is the first step in order to fully remove bi-directional search. 
Bi-directional search is being removed because after recent changes, 
bi-directional search will be SIGNIFICANTLY slower than traditional 
backward search, take more memory, and also is also unable to be 
thread-parallelized as efficiently. It was originally included because 
HMMER uses it, but now AwFmIndex using backward-only will be 
significantly faster than the current bi-directional in hmmer.