change kmer table func to return value, not pointer.
This might change back later, but I think I can use this implementation to actually implement sub-kmer memoization. 
to perform sub-kmer memoization, imagine searching for GAGC in a 10-mer table. the beginning of the GAGA's is one after the end of the GAGATTTTTT element, and includes everything up to one less than the GAGTAAAAAA start ptr.