fix bugs where prefix sums weren't being written, fileHandle wasn't set
Instead of giving fwrite the array of prefix sums to write, it was given the address of the prefixSums pointer in the index struct. 

Also, the fileHandle wasn't getting set on file read, which causes the file to never be closed, and to cause garbage when the index was being deallocated.