change file mode to allow for reading seq/SA, fix file offset functions.
Before, when the index was written to file, it only had write permissions. Because of this, reading the suffix array or sequence would work when the index was read off disk, but not when it was intially created. 

Also, some of the values in awFmGetSequenceFileOffset were incorrect, they should be right this time.