fix bug in fasta parsing, print out sequence part for debug.
This commit fixes a bug in this example where newlines were being added to the index. The new code checks the last char from the fgets(), and if it's a newline, the newline is eliminated.


For debug purposes, the sequence (or a section at the beginning and end) will be printed to make sure everything looks correct.