update index allocation to use malloc, except where alignment guaranteed
Unfortunately, aligned_alloc reqires the allocated size to be a multiple of the alignment width, so it's not really possible to do things like the index struct and prefix sums. Therefore, these have been changed to malloc calls, but the BWT should still be aligned nicely.

good chatch, address sanitizer!