change "AwFmMetadata" to "AwFmConfiguration" to show its a user config
Metadata was a little ambiguous, so this change makes it obvious what is 
a user-configuratble parameter, and what is not. In this change, version 
was taken out of the user-configurable settings, and moved into the 
creation, where it can be checked without having to (reasonably) worry 
if the user messed with it.