Fix remaining issue with long kmers
This commit further fixes an issue that was partially fixed in the previous version. In places like the kmer seed table, uint8_t's were still used for kmer length and position variables. Now size_t's are used for all length and positions, meaning that all lengths of kmers will be supported.